+++ Libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: itrueba- <itrueba-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/30 11:08:55 by itrueba-          #+#    #+#             */
-/*   Updated: 2023/02/01 19:02:14 by itrueba-         ###   ########.fr       */
+/*   Updated: 2023/02/10 12:08:41 by itrueba-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,7 @@ typedef struct s_list
 void				ft_bzero(void *s, size_t n);
 int					ft_isdigit(int c);
 void				*ft_memset(void *b, int c, size_t len);
-int					ft_atoi(const char *nptr);
+long long int		ft_atoi(const char *nptr);
 void				*ft_calloc(size_t nmemb, size_t size);
 t_list				*ft_lstnew(int content);
 t_list				*ft_pre_lstlast(t_list *lst);
